mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "altair==5.5.0",
    "fastmcp>=2.12.4",
    "langchain>=1.1.0",
    "langchain-mcp-adapters>=0.1.14",
    "langchain-openai>=1.1.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
